Move validate_and_return_updater_request to utils
This is a general function to check updated_by for all write requests.
