+++ app/main/views/users.py
@@ -8,9 +8,8 @@ from .. import main
 from ... import db, encryption
 from ...models import User, AuditEvent, Supplier, Framework, SupplierFramework, DraftService
 from ...utils import get_json_from_request, json_has_required_keys, \
-    json_has_matching_id, pagination_links, get_valid_page_or_1
-from ...service_utils import validate_and_return_updater_request
-from ...validation import validate_user_json_or_400, validate_user_auth_json_or_400, is_valid_buyer_email
+    json_has_matching_id, pagination_links, get_valid_page_or_1, validate_and_return_updater_request
+from ...validation import validate_user_json_or_400, validate_user_auth_json_or_400
 
 
 @main.route('/users/auth', methods=['POST'])
