+++ app/utils.py
@@ -3,6 +3,15 @@ from flask import abort, request
 from six import iteritems, string_types
 from werkzeug.exceptions import BadRequest
 
+from .validation import validate_updater_json_or_400
+
+
+def validate_and_return_updater_request():
+    json_payload = get_json_from_request()
+    json_has_required_keys(json_payload, ['update_details'])
+    validate_updater_json_or_400(json_payload['update_details'])
+    return json_payload['update_details']
+
 
 def link(rel, href):
     """Generate a link dict from a rel, href pair."""
